added a progressBar, fixed a wrong cast of duration to int

+++ app/src/main/java/com/slidetimer/oli/slidetimer/SlideListActivity.java
@@ -38,7 +38,7 @@ public class SlideListActivity extends AppCompatActivity implements View.OnClick
 
         //load previous user inputs
         dataFromMain = getIntent().getExtras();
-        int duration = dataFromMain.getInt("duration");
+        double duration = dataFromMain.getDouble("duration");
         int numOfSlides = dataFromMain.getInt("numOfSlides");
         String name = dataFromMain.getString("name");
